Simplify the calling for BoxPlotter
